requirements.txt
-e .
pytest
aws_cdk.aws_lambda_event_sources
aws_cdk.aws_athena
aws_cdk.aws_stepfunctions
aws_cdk.aws_stepfunctions_tasks
aws_cdk.aws_lambda_python